+++ app.vue
@@ -13,13 +13,7 @@ import {
   PushNotifications,
   Token,
 } from '@capacitor/push-notifications'
-import {
-  useSettingSideBar,
-  useLoginSideBar,
-  useSigninSideBar,
-  useIsApp,
-  useCurrentUserProfile,
-} from '~/composables/states'
+import { useIsApp, useCurrentUserProfile } from '~/composables/states'
 import {
   useBrowserTopColor,
   useBrowserTopColorDarkMode,
@@ -39,10 +33,6 @@ const browserTopColorDarkMode = useBrowserTopColorDarkMode()
 
 const isRefreshing = shallowRef(false)
 const acceptNotifications = shallowRef(false)
-
-const settingsSideBar = useSettingSideBar()
-const loginSideBar = useLoginSideBar()
-const signinSideBar = useSigninSideBar()
 
 const isApp = useIsApp()
 
@@ -345,66 +335,7 @@ useHead({
   </NuxtLayout>
   <Toast />
   <AudioPlayer />
-  <Sidebar
-    v-model:visible="settingsSideBar"
-    :baseZIndex="10000"
-    position="right"
-    class="w-full"
-    blockScroll
-    id="settings-sidebar"
-    @hide="
-      () => {
-        trackClickEvent(
-          'Click Tracking - Settings Sidebar Close Button',
-          'Settings Sidebar',
-          `close sidebar`
-        )
-      }
-    "
-  >
-    <template #header><h1 class="font-medium">Settings</h1></template>
-    <Settings />
-  </Sidebar>
-  <Sidebar
-    v-model:visible="loginSideBar"
-    :baseZIndex="10001"
-    position="right"
-    class="w-full"
-    blockScroll
-    id="login-sidebar"
-    @hide="
-      () => {
-        trackClickEvent(
-          'Click Tracking - Login Sidebar Close Button',
-          'Login Sidebar',
-          `close sidebar`
-        )
-      }
-    "
-  >
-    <template #header><h1 class="font-medium">Log in</h1></template>
-    <LogIn />
-  </Sidebar>
-  <Sidebar
-    v-model:visible="signinSideBar"
-    :baseZIndex="10001"
-    position="right"
-    class="w-full"
-    blockScroll
-    id="signin-sidebar"
-    @hide="
-      () => {
-        trackClickEvent(
-          'Click Tracking - Sign in Sidebar Close Button',
-          'Sign in Sidebar',
-          `close sidebar`
-        )
-      }
-    "
-  >
-    <template #header><h1 class="font-medium">Sign in</h1></template>
-    <SignIn />
-  </Sidebar>
+  <Sidebars />
 </template>
 
 <style lang="scss">
