+++ composables/states.ts
@@ -22,6 +22,13 @@ export const useCurrentUserProfile = () => useState('useCurrentUserProfile', () 
 // setting sidebar state
 export const useSettingSideBar = () => useState('useSettingSideBar', () => false)
 
+// login sidebar state
+export const useLoginSideBar = () => useState('useLoginSideBar', () => false)
+
+// signin sidebar state
+export const useSignupSideBar = () => useState('useSignupSideBar', () => false)
+
+
 // global state if this instance is a native app
 export const useIsApp = () => useState('useIsApp', () => false)
